requirements.txt
Flask
Flask_SQLAlchemy
python-dotenv
openai
